requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
discord.py==1.6.0
idna==3.1
logging==0.4.9.6
multidict==5.1.0
psycopg2==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyxivapi==0.5.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.4
websockets==8.1
yarl==1.6.3
